Non-implicit conversion from Vector2 to Vector3.
It shouldn't be implicit, because then it will be possible to
autoconvert Vector4 from Vector2, which shouldn't be possible at all.
But it shouldn't be explicit either, because this will not be possible
then:

    Vector2 vec2;
    Vector3 vec3 = {vec2, 1.0};

Vector4 constructor from Vector3 stays the same, because the conversion
is fairly common, nearly always with W set to 1.