Absolute include paths in documentation.
Currently it was incorrectly stated that for using Math::Matrix or for
example Trade::MeshData it was sufficient to

    #include <Matrix.h>
    #include <MeshData.h>

which was obviously wrong, as Math/ (nor any other namespace-related
subdirectory) is not on include path. Now it states correctly that a
directory prefix is needed:

    #include <Math/Matrix.h>
    #include <Trade/MeshData.h>

The same is now fixed for classes in magnum-plugins repository.