Deprecating AbstractShaderProgram::bind*Location().
Preferred workflow is to specify attribute location explicitly in the
shader. The functions remains here as some old Intel systems don't
support the required extension ARB_explicit_attrib_location (and it's
not in GL 3.0 either). Also updated and fixed the documentation.

Function bindAttribute() was renamed to bindAttributeLocation() to be
consistent with bindFragmentDataLocation().

PhongShader now uses explicit attribute location.