Another, more convenient way to swizzle.
On the other hand everything of this is done at runtime, so it's less
performant than the previous version, mainly when used in loops. When
the result is declared as constexpr, it is done at compile time, just
like the previous version.

I don't know which version to keep, so there will be both until a good
decision.