Added NumericType and FloatingPointType to Math::MathTypeTraits.
NumericType is corresponding numeric type with size at least the same as
int. It is used in debug operators for Matrix and Vector to prevent
printing chars as characters.

If the type isn't already floating-point, FloatingPointType is
corresponding larger type with sufficient size for normalization of
given integral type.

Also updated type traits for long types, they are now subclassed either
from int or long long based on sizeof(long).