Using specialized Vector type in images and textures.
Thanks to DimensionTraits it is now possible to e.g. conveniently access
components by name or pass size as combination of vector and scalar:

    GLsizei width = image.size().x();
    image.setData({xy, 1}, ...);

Instead of previous inconvenient ways:

    GLsizei width = image.size()[0];
    Math::Vector2<GLsizei> size(xy, 1);
    image.setData(size, ...);

Not using the specialized type for internal functions and storage, as it
wouldn't cause any other improvements. This way it is virtually possible
to forward-declare the specialized types without including them in the
headers.