Using EXT_direct_state_access/ARB_separate_shader_objects for uniforms.
Also explicitly calling use() on default (non-DSA) setUniform()
implementation. Because of that, it is now possible to conveniently call
use() at the end, instead of at the beginning of uniform setting chain,
for example:

    // before
    shader->use();
    shader->setTransformation(transformation)
        ->setProjection(projection)
        ->setColor(color);

    // now
    shader->setTransformation(transformation)
        ->setProjection(projection)
        ->setColor(color)
        ->use();

Note that you still have to explicitly call use(), because if DSA is
available, no use() is called explicitly on any setUniform(). If DSA is
not available, use() is called on first setUniform() and subsequent
calls have no negative performance impacts.