SceneGraph: new Animable feature.
Allows to animate objects, pause, repeat and resume the animations. The
user implements animation step and can perform actions on state changes.

Each Animable is part of some group which is optimized for case when no
animation is running, thus it is possible to create multiple independent
"animation islands" to improve performance.