Math: removed integer/float RectangularMatrix multiplication and division.
We would need to duplicate all the functionality found in Vector, which
I don't think is needed at all. If anyone needs to do this, it is
possible to "linearize" the matrix into long vector and do the
operations on it.

Also updated subclass operator implementation, added tests for it, both
for proper returned value and proper result type.