MeshTools: return Containers::Array instead of pointer from compressIndices().
This is source-incompatible change, but this function probably wasn't
used much in user code, rather its Mesh/Buffer convenience overload.

Actually Containers::Array<char> has _implicit_ conversion operator to
char*, but this should not be an issue, as the conversion is not allowed
on rvalue references, thus the following code shouldn't compile at all.
Not available on GCC < 4.8.1, though.

        // ...
        char* data;
        std::tie(..., data) = MeshTools::compressIndices(...);