Using Range2D instead of Rectangle everywhere.
Encourages vectorization and generic usage even more. Some functions
were rewritten to make use of the new features, resulting in shorter and
more readable code. This also fixes the annoying naming collision with
WINAPI Rectangle() function.

The old Rectangle is now subclass of Range2D, is marked as deprecated
and will be removed in future release.