Bumped minimal version of exts related to layout() GLSL qualifier to 1.50.
Related to issue #27 -- layout() qualifier causes compiler errors on
GLSL 1.40 (GL 3.1) on Mac OS X even that the extensions are marked as
supported. Bumping the minimal required version disables the extensions
on this versions, which should avoid the issue.

It is not proper fix at all, but most of the GL3 implementations are
now at least GL 3.2, so it shouldn't be too much of an issue. The only
exception is Mesa (GL 2.1/3.1 on most of the machines), but we need
to maintain GL 2.1 compatibility for it anyway.