MeshTools: reworked combineIndexedArrays().
Added combineIndexArrays() function, which can combine any number of
indexarrays (i.e. number specified at runtime). The function originated
in ColladaImporter plugin and is moved here to make it useful elsewhere.
It is now used as base for combineIndexedArrays(), which is further
simplified without unnecessary functor-like implementation. Also
improved documentation.

The combineIndexedArrays() function now takes std::pair instead of
std::tuple, the previous version is alias to the new one, is marked as
deprecated and will be removed in some future release.