Explicitly passing shader to Mesh*::draw().
Previously the API didn't encourage the user to set up and activate
shader before drawing the meshes, leading to unintuitive behavior:

    // Can I just call draw() or do I have to fully understand the
    // meaning of the universe before?
    mesh.draw();

Now the draw() needs the shader passed explicitly as parameter, which
should hint that the shader must be set up somehow:

    // Right, so this needs just a shader and that's all. Expecting this
    // I fortunately *did* configure all the uniforms before this call.
    mesh.draw(shader);

It is also possible to pass the shader as rvalue, in case the drawing is
just a one-off thing and is already fully configured.

    mesh.draw(MyShader{});

As usual, the original API is kept, is marked as deprecated and will be
removed in some future release.