Full ARB_direct_state_access support in textures.
ARB_DSA is now preferred in single-bind cases, as it is easier to use
than passing pointers to ARB_multi_bind. ARB_multi_bind was preferred
for single-bind previously simply because EXT_DSA was not in core.

Because there is a lot to say about feature selection for each function,
I took this as a opportunity to remove redundant documentation blocks,
just refer to Texture documentation from everywhere and add extension
requirements and deprecation where needed, so it's clear for each class
what needs what.

ARB_DSA also took the opportunity to finally remove all target enum
values from function calls and because of that the CubeMapTexture has to
handle a bunch of special cases. In order:

-   CubeMapTexture::imageSize() now doesn't take face parameter and
    returns one value for all faces. I'm thus now also assuming that the
    user is sane and called either setStorage() or setImage() with the
    same size for all faces. In non-ARB_DSA path I'm thus querying only
    size of +X face and returning it as size for all faces. The old
    imageSize(Coordinate, Int) overload is still present, but ignores
    the first parameter and calls imageSize(Int). It is marked as
    deprecated and will be removed in some future release.
-   CubeMapTexture::image() now needs to call glTextureSubImage() in
    ARB_DSA path to make it possible to extract single face. Other code
    paths (EXT_DSA, Robustness and "default") remain the same.
-   CubeMapTexture::setSubImage() calls glTextureSubImage3D() in
    ARB_DSA, because it is not possible to specify face index in
    glTextureSubImage2D(). Other code paths (EXT_DSA and "default")
    remain the same.

Implementation of these special cases is extracted into CubeMapTexture
class to avoid pollution of AbstractTexture with incompatible nonsense.