Math: made dot(), angle(), *lerp() and cross() free functions.
It is often annoying to write e.g. this, especially in generic code:

    T dot = Math::Vector<size, T>::dot(a, b);

When this is more than enough and the compiler can infer the rest from
the context:

    T dot = Math::dot(a, b);

There are more downsides and confusing cases (you can call
Math::Vector<3, T>::dot(), Math::Vector3<T>::dot() and Color3::dot() and
it is still the same function), so I made these as free functions in
Math namespace. You can now also abuse ADL for the calls, but I would
advise against that for better readability:

    T d = dot(a, b); // dot?! what on earth is dot? and what is a?

The only downside found when porting is that you need to specify the
type somehow when having both parameters as initializer lists:

    T d = dot({2.0f, -1.5f}, {1.0f, 2.5f});        // error
    T d = dot(Complex{2.0f, -1.5f}, {1.0f, 2.5f}); // okay

But that's probably reasonable (and it's also highly corner case,
the functions were used this way only in tests).

The original static member functions are of course still present, but
marked as deprecated and will be removed at some point in future.