Better fix to ensure a single static ResourceManager instance.
I'm not exactly sure where was the actual problem, but it seems to
be related to definition of templated function with static member
variables in header file in combination with dynamic libraries, causing
multiple definitions of a symbol which did or did not lead to compile
error and/or things breaking in horribly unpredictable way.

Previously it was done via (undocumented) macros, depending on include
order, which is great until it breaks horribly after random unrelated
changes. Currently it is done via black template magic and I'm
still afraid of including it in the public documentation.

There should be no visible change to the end user except for possibly
one less randomly weird bug (#79? couldn't reproduce).

... well, looking at the diff, the solution is still an ugly **********,
but at least it seems to work.