Simplify Mesh class internals.
The original goal was to avoid branches when binding the vertex
attributes for drawing, so I stored float, integral and double
attributes in separate std::vector instances and then was going through
each one of them in separate loop. In retrospect that was _not_ a good
idea, because it results in larger Mesh class, two more allocations
resulting in far more pointer chasing and more complicated
constructor/destructor.

Now everything is stored in a single vector. I may optimize it further
by not calling the constructor/destructor on it when VAOs are used.