Math: more explicit default zero/identity constructors.
Some classes are by default constructed zero-filled while other are set
to identity and the only way to to check this is to look into the
documentation. This changes the default constructor of all classes to
take an optional "tag" which acts as documentation about how the type is
constructed. Note that this result in no behavioral changes, just
ability to be more explicit when writing the code. Example:

    // These two are equivalent
    Quaternion q1;
    Quaternion q2{Math::IdentityInit};

    // These two are equivalent
    Vector4 vec1;
    Vector4 vec2{Math::ZeroInit};

    Matrix4 a{Math::IdentityInit, 2}; // 2 on diagonal
    Matrix4 b{Math::ZeroInit};        // all zero

This functionality was already present in some ugly form in Matrix,
Matrix3 and Matrix4 classes. It was long and ugly to write, so it is
now generalized into the new Math::IdentityInit and Math::ZeroInit tags,
the original Matrix::IdentityType, Matrix::Identity, Matrix::ZeroType
and Matrix::Zero are deprecated and will be removed in the future
release.

    Math::Matrix<7, Int> m{Math::Matrix<7, Int>::Identity}; // before
    Math::Matrix<7, Int> m{Math::IdentityInit};             // now